+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial release. Support for text generation via Gemini Nano on Android.